move event file
