+++ packages/contracts/test/test-utils/plugin-setup-processor.ts
@@ -1,5 +1,5 @@
 import {ethers} from 'hardhat';
-import {findEvent} from './event';
+import {findEvent} from '../../utils/event';
 import {PluginSetupProcessor, PluginRepoRegistry} from '../../typechain';
 
 export async function deployPluginSetupProcessor(
