+++ packages/contracts/tsconfig.json
@@ -7,8 +7,8 @@
     "outDir": "dist",
     "declaration": true,
     "moduleResolution": "node",
-    "resolveJsonModule": true,
+    "resolveJsonModule": true
   },
-  "include": ["./scripts", "./test", "./typechain"],
+  "include": ["./scripts", "./test", "./typechain", "utils/event.ts"],
   "files": ["./hardhat.config.ts"]
 }
